pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-notes-assistant"
version = "0.1.0"
description = "A RAG server for your Obsidian vault."
requires-python = ">=3.12"
dependencies = [
    "python-dotenv",
    "openai",
    "ollama",
    "torch",
    "pymilvus",
    "sentence-transformers",
    "watchdog",
    "PyYAML",
    "mcp",
    "colorlog"
]

[project.scripts]
run-obsidian-rag = "main:main"

[tool.setuptools.packages.find]
where = ["."]